requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
itsdangerous==2.1.2
email-validator==2.1.0
